Remove cf code and reduce api calls per space
There was some CF specific code that was moved to the go cfclient
library via
https://github.com/cloudfoundry-community/go-cfclient/pull/142
for buildpack info.

As a result, we could remove that buildpack code from this repo and
updated our lock file to use it.

In addition, we made two API calls per space. one for space managers and
one for space developers. Instead, there's a call to get all the roles
for the space. We moved to that 1 call and had to add a little bit of
filtering logic to search for our specific role now.

Tests updated appropriately